csv-uploader_app/requirements.txt
# Core Dash dependencies
dash==2.17.1
dash-bootstrap-components==1.5.0
plotly==5.17.0

# Data processing
pandas==2.1.4
numpy==1.24.4

# Databricks integration
databricks-sdk==0.18.0
databricks-sql-connector==3.0.2

# HTTP requests for API calls
requests==2.31.0

# Environment configuration
python-dotenv==1.0.0

# Optional development dependencies
black==23.12.1
flake8==7.0.0
pytest==7.4.4